improvment for ExtraInfoView
improve the preparation time label size to it fits the entire text it has to display. Also trying to display preparation time for every recipes (shown in console but not on screen everyTime, have to work on where the refreshData and configure view method are called in recipeCell for ExtraInfoView).